+++ front/change_supplier.form.php
@@ -67,8 +67,6 @@ if (isset($_POST["update"])) {
         sprintf(__('%s deletes an actor'), $_SESSION["glpiname"])
     );
     Html::redirect(Change::getFormURLWithID($link->fields['changes_id']));
-} else if (isset($_GET["id"])) {
-    $link->showSupplierNotificationForm($_GET["id"]);
 } else {
     Html::displayErrorAndDie('Lost');
 }