+++ front/ticket_user.form.php
@@ -80,8 +80,6 @@ if (isset($_POST["update"])) {
     );
 
     Html::redirect($CFG_GLPI["root_doc"] . "/front/ticket.php");
-} else if (isset($_GET["id"])) {
-    $link->showUserNotificationForm($_GET["id"]);
 } else {
     Html::displayErrorAndDie('Lost');
 }